Fix connection error for code collaboration PORT 1234
